1) Added "Ease of Movement" (EMV) trading strategy (23 strategies total now)
2)
We no longer track the reasons why we are buying.
Each strategy has 1 scenario at which we buy so there is no reason to track.

We will still track the reasons to sell, but there are only 2
Each strategy has 1 scenario at which we sell, and then we have logic to sell at hard stop, so there are 2 reasons